+++ src/distilltwin/__init__.py
@@ -1,5 +1,6 @@
 """DistillTwin: an open, control-oriented distillation digital twin."""
 
+from distilltwin.estimation import ExtendedKalmanFilter, MeasurementConfig, MeasurementModel
 from distilltwin.model import ColumnConfig, ColumnInputs, DistillationColumn
 from distilltwin.scenarios import Scenario, ScenarioRunner
 
@@ -7,6 +8,9 @@ __all__ = [
     "ColumnConfig",
     "ColumnInputs",
     "DistillationColumn",
+    "ExtendedKalmanFilter",
+    "MeasurementConfig",
+    "MeasurementModel",
     "Scenario",
     "ScenarioRunner",
 ]